# Conflicts:
#	iOrder_iOS/iOrder.xcodeproj/project.pbxproj
#	iOrder_iOS/iOrder/Classes/MainModule/Home/Controller/IODishViewController.m

+++ iOrder_iOS/iOrder/Classes/Other/Utilities/Dishes/Param/YWJDishesParam.h
@@ -13,6 +13,6 @@
 /**
  *  商铺ID，唯一标识
  */
-@property (nonatomic, assign) int shopId;
+@property (nonatomic, assign) NSInteger shopId;
 
 @end
